Move format_article_title to a new class
Having it Utils caused a conflict when I tried to use it in a controller.
