+++ app/models/assignment.rb
@@ -13,7 +13,7 @@
 #  wiki_id       :integer
 #
 
-require "#{Rails.root}/lib/utils"
+require "#{Rails.root}/lib/article_utils"
 
 #= Assignment model
 class Assignment < ActiveRecord::Base
@@ -54,7 +54,7 @@ class Assignment < ActiveRecord::Base
   private
 
   def set_defaults_and_normalize
-    self.article_title = Utils.format_article_title(article_title) unless article_title.nil?
+    self.article_title = ArticleUtils.format_article_title(article_title) unless article_title.nil?
     self.wiki_id ||= course.home_wiki.id
   end
 end
