+++ lib/utils.rb
@@ -27,14 +27,4 @@ class Utils
       (array || model.current).each(&method)
     end
   end
-
-  # This method takes user input and tries to convert it into a valid article title
-  def self.format_article_title(article_title)
-    formatted_title = article_title.clone
-    first_letter = formatted_title.chars.first
-    # Use mb_chars so that we can capitalize unicode letters too.
-    formatted_title[0] = first_letter.mb_chars.capitalize.to_s
-    formatted_title = formatted_title.tr(' ', '_')
-    formatted_title
-  end
 end
